+++ src/visualizations/config/adapters/dhis_highcharts/noData.js
@@ -1,0 +1,8 @@
+export default function() {
+    return {
+        style: {
+            fontSize: '13px',
+            fontWeight: 'normal',
+        },
+    }
+}