+++ src/visualizations/config/adapters/dhis_highcharts/subtitle/gauge.js
@@ -1,0 +1,11 @@
+import isString from 'd2-utilizr/lib/isString'
+
+export default function(series, layout, dashboard, filterTitle) {
+    const seriesName = series[0].name
+
+    const mergedTitle = seriesName + ' - ' + filterTitle
+
+    return {
+        text: dashboard || isString(layout.title) ? mergedTitle : seriesName,
+    }
+}