+++ src/visualizations/store/adapters/dhis_dhis/singleValue.js
@@ -1,0 +1,5 @@
+export default function(acc, seriesIds, categoryIds, idValueMap, metaData) {
+    const seriesId = seriesIds[0]
+
+    acc.push(idValueMap.get(seriesId))
+}