+++ src/visualizations/store/adapters/dhis_highcharts/gauge.js
@@ -1,0 +1,7 @@
+export default function(acc, seriesIds, categoryIds, idValueMap, metaData) {
+    const seriesId = seriesIds[0]
+
+    acc.push({
+        data: [parseFloat(idValueMap.get(seriesId))],
+    })
+}