+++ src/visualizations/store/adapters/dhis_highcharts/pie.js
@@ -1,0 +1,8 @@
+export default function(acc, seriesIds, categoryIds, idValueMap, metaData) {
+    acc.push(
+        ...seriesIds.map(seriesId => ({
+            name: metaData.items[seriesId].name,
+            y: parseFloat(idValueMap.get(seriesId)),
+        }))
+    )
+}