+++ src/visualizations/store/adapters/index.js
@@ -1,0 +1,7 @@
+import dhis_highcharts from './dhis_highcharts'
+import dhis_dhis from './dhis_dhis'
+
+export default {
+    dhis_highcharts,
+    dhis_dhis,
+}